# docs/swagger.yaml
definitions:
...
  api.ReportForSwagger:
    properties:
      fueling_records:
        items:
          $ref: '#/definitions/api.myFuelingRecord'
        type: array
      maintenance_records:
        items:
          $ref: '#/definitions/api.myMaintenanceRecord'
        type: array
      vehicle:
        $ref: '#/definitions/api.createVehicleResponse'
    type: object
...
  api.createMaintenanceRecordRequest:
    properties:
      maintenance_date:
        type: string
      maintenance_person_id:
        type: integer
      mileage_at_service:
        type: integer
      notes:
        type: string
      service_type:
        type: string
      status:
        type: string
      total_cost:
        type: number
      vehicle_id:
        type: integer
    type: object
  api.createMaintenanceRecordResponse:
    properties:
      maintenance_date:
        type: string
      maintenance_person_id:
        type: integer
      mileage_at_service:
        type: integer
      notes:
        type: string
      service_type:
        type: string
      status:
        type: string
      total_cost:
        type: number
      vehicle_id:
        type: integer
    type: object
  api.createTaskRequest:
    properties:
      driver_id:
        type: integer
      end_latitude:
        type: number
      end_longitude:
        type: number
      end_time:
        type: string
      notes:
        type: string
      start_latitude:
        type: number
      start_longitude:
        type: number
      start_time:
        type: string
      status:
        type: string
    type: object
  api.createTaskResponse:
    properties:
      ID:
        type: integer
      driver_id:
        type: integer
      end_latitude:
        type: number
      end_longitude:
        type: number
      end_time:
        type: string
      notes:
        type: string
      start_latitude:
        type: number
      start_longitude:
        type: number
      start_time:
        type: string
      status:
        type: string
    type: object
...
  api.createVehicleResponse:
    properties:
      ID:
        type: integer
      assigned_driver:
        type: integer
      car_model:
        type: string
      color:
        type: string
      current_mileage:
        type: integer
      last_maintenance:
        type: string
      license_plate:
        type: string
      make:
        type: string
      next_maintenance:
        type: string
      notes:
        type: string
      sitting_capacity:
        type: integer
      status:
        type: string
      type:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
  api.myFuelingRecord:
    properties:
      after_fueling_image:
        type: string
      amount:
        type: number
      before_fueling_image:
        type: string
      fueling_person_id:
        type: integer
      gas_station:
        type: string
      notes:
        type: string
      total_cost:
        type: number
      vehicle_id:
        type: integer
    type: object
  api.myMaintenanceRecord:
    properties:
      maintenance_date:
        type: string
      maintenance_person_id:
        type: integer
      mileage_at_service:
        type: integer
      notes:
        type: string
      service_type:
        type: string
      status:
        type: string
      total_cost:
        description: Parts               []Part             `gorm:"foreignKey:ID"
          json:"parts"`
        type: number
      vehicle_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This API serves as a backend for Vehicle Management System
  title: Vehicle Management System API
  version: "1.0"
paths:
...
  /maintenance:
    get:
      consumes:
      - application/json
      description: Get all maintenance records if user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.createMaintenanceRecordResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all maintenance records of particular user
      tags:
      - maintenance
    post:
      consumes:
      - application/json
      description: Create a maintenance record
      parameters:
      - description: Maintenance
        in: body
        name: maintenance
        required: true
        schema:
          $ref: '#/definitions/api.createMaintenanceRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createMaintenanceRecordResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a maintenance record
      tags:
      - maintenance
  /maintenance/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a maintenance record
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createMaintenanceRecordResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a maintenance record
      tags:
      - maintenance
    get:
      consumes:
      - application/json
      description: Get a maintenance record
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createMaintenanceRecordResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a maintenance record
      tags:
      - maintenance
    put:
      consumes:
      - application/json
      description: Update a maintenance record
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Maintenance
        in: body
        name: maintenance
        required: true
        schema:
          $ref: '#/definitions/api.createMaintenanceRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createMaintenanceRecordResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a maintenance record
      tags:
      - maintenance
  /report/{vehicle_id}:
    get:
      consumes:
      - application/json
      description: Get a report
      parameters:
      - description: Vehicle ID
        in: path
        name: vehicle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReportForSwagger'
      security:
      - ApiKeyAuth: []
      summary: Get a report
      tags:
      - report
  /task:
    get:
      consumes:
      - application/json
      description: Get all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/api.createTaskResponse'
              type: array
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all tasks
      tags:
      - task
    post:
      consumes:
      - application/json
      description: Create a task
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/api.createTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createTaskResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a task
      tags:
      - task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createTaskResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a task
      tags:
      - task
    get:
      consumes:
      - application/json
      description: Get a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createTaskResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a task
      tags:
      - task
    put:
      consumes:
      - application/json
      description: Update a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/api.createTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createTaskResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a task
      tags:
      - task
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"